refactor: move extracted components to features
